# shopsphere/services/catalog-service/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: ShopSphere Catalog Service API
  version: "1.0"

servers:
  - url: http://localhost:3004/api/v1

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /products:
    get:
      summary: List products (with HATEOAS)
      responses:
        "200":
          description: OK
    post:
      summary: Create product (ADMIN only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - categoryId
                - brand
                - inStock
              properties:
                name:
                  type: string
                  example: Perfume Z
                price:
                  type: number
                  example: 120
                categoryId:
                  type: integer
                  example: 1
                brand:
                  type: string
                  example: BrandZ
                inStock:
                  type: boolean
                  example: true
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "401":
          description: Unauthorized (missing/invalid token)
        "403":
          description: Forbidden (not admin)

  /products/{id}:
    get:
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not found

    put:
      summary: Update product (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Perfume Updated
                price:
                  type: number
                  example: 150
                categoryId:
                  type: integer
                  example: 1
                brand:
                  type: string
                  example: BrandUpdated
                inStock:
                  type: boolean
                  example: false
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)
        "404":
          description: Not found

    delete:
      summary: Delete product (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)
        "404":
          description: Not found

  /categories:
    get:
      summary: List categories
      responses:
        "200":
          description: OK

    post:
      summary: Create category (ADMIN only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Perfumes
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)

  /categories/{id}:
    get:
      summary: Get category by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not found

    put:
      summary: Update category (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Updated Category
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)
        "404":
          description: Not found

    delete:
      summary: Delete category (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)
        "404":
          description: Not found

  /solr/reindex:
    post:
      summary: Reindex products from Postgres into Apache Solr
      responses:
        "200":
          description: Indexed successfully
        "502":
          description: Solr indexing error

  /search/products:
    get:
      summary: Search products using Apache Solr index (advanced queries)
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          example: '(brand:"BrandX" OR brand:"Guess Brand") AND price:[50 TO 120]'
        - in: query
          name: category
          required: false
          schema:
            type: integer
          example: 1
        - in: query
          name: brand
          required: false
          schema:
            type: string
          example: BrandX
        - in: query
          name: inStock
          required: false
          schema:
            type: boolean
          example: true
        - in: query
          name: minPrice
          required: false
          schema:
            type: number
          example: 50
        - in: query
          name: maxPrice
          required: false
          schema:
            type: number
          example: 120
        - in: query
          name: sort
          required: false
          schema:
            type: string
          example: "price asc"
        - in: query
          name: page
          required: false
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Solr search results
        "500":
          description: Solr query failed